backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
bcrypt
PyJWT
python-multipart